+++ gta5-mod/ThirdSpaceGTAV/Properties/AssemblyInfo.cs
@@ -1,0 +1,19 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("ThirdSpaceGTAV")]
+[assembly: AssemblyDescription("Third Space Vest haptic integration for GTA V")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Third Space")]
+[assembly: AssemblyProduct("ThirdSpaceGTAV")]
+[assembly: AssemblyCopyright("Copyright © 2024")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(false)]
+
+[assembly: Guid("c1d2e3f4-a5b6-7890-cdef-123456789012")]
+
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
+